+++ src/Payment/resources/manifest.php
@@ -4,5 +4,5 @@ declare(strict_types=1);
 
 return [
     'name' => 'Vanilo Payment Module',
-    'version' => '5.1-dev',
+    'version' => '5.1.0',
 ];