+++ src/Product/Changelog.md
@@ -6,7 +6,7 @@
 ##### 2025-XX-XX
 
 - Dropped PHP 8.2 Support
-- Dropped Laravel 10 Support
+- ~~Dropped Laravel 10 Support~~
 - Added Laravel 12 Support
 - Added the `gtin` field to the products and master product variants tables